+++ Changelog.md
@@ -1,3 +1,10 @@
+# v0.2.0
+
+## Breaking Changes!
+
+- Rename `RtsObj` component to `Obstacle`
+- Rename `BoidsUpdater` resource to `BoidUpdater`
+
 # v0.1.0
 
 ## Initial Release!